fix(update-doc): move createDocBlocks to dev schema, fix codegen
createDocBlocks and its associated types (CreateBlockInput, TextBlock,
BlockAlignment, etc.) only exist in the dev schema but were defined in
.graphql.ts files validated against the production schema, causing codegen
to fail with 10 errors.

- Move createDocBlocks mutation to update-doc-tool.graphql.dev.ts
- Import CreateDocBlocksMutation/Variables from graphql.dev/graphql
- Import all dev-only types in helpers from graphql.dev/graphql
- Add versionOverride: 'dev' to the createDocBlocks API request
- Regenerate both graphql/ and graphql.dev/ type files
